correct a stupid typo
one which we should fix with #200
